MAJOR Add nested Nav and remove CatalogPage
Add getProductCategoryNav and getProductCategories methods to controller extension.
This allows either a data list of Categories or a fully rendered nav of categoires to be generated.

Remove CatalogPage, as product cataloges don't need to mix with Pages, it is cleaner if the two are independandt.
This means that now there is just a cataloge controller extending page controller (to inherit methods).